snapped_intersect_ray update
added snapped_intersect_ray function that uses raytraces parallel to the ground to reduce step_up calls. 
step_up has also been improved for more accurate usage. It now returns a dictionary with the collision point and collision normal rather than just a location to move to. This was done to make the returned results more clear and usage of the results in other scripts less constrained to a specific implementation by returning general data.